Move progress tracking into own module.
