+++ data/src/lib.rs
@@ -8,7 +8,6 @@ mod file_cleaner;
 mod file_downloader;
 mod file_upload_session;
 pub mod migration_tool;
-mod progress_tracking;
 mod prometheus_metrics;
 mod remote_client_interface;
 mod repo_salt;
@@ -21,5 +20,4 @@ pub use cas_client::CacheConfig;
 pub use deduplication::RawXorbData;
 pub use file_downloader::FileDownloader;
 pub use file_upload_session::FileUploadSession;
-pub use progress_tracking::CompletionTracker;
 pub use xet_file::XetFileInfo;
