+++ deduplication/Cargo.toml
@@ -8,6 +8,7 @@ gearhash = "0.1.3"
 merklehash = { path = "../merklehash" }
 mdb_shard = { path = "../mdb_shard" }
 merkledb = { path = "../merkledb" }
+progress_tracking = { path = "../progress_tracking" }
 utils = { path = "../utils" }
 more-asserts = "0.3.1"
 async-trait = "0.1.53"
