+++ progress_tracking/src/upload_tracking.rs
@@ -2,11 +2,18 @@ use std::collections::{BTreeSet, HashMap};
 use std::mem::take;
 use std::sync::Arc;
 
-use deduplication::FileXorbDependency;
 use merklehash::MerkleHash;
 use more_asserts::debug_assert_le;
 use tokio::sync::Mutex;
-use utils::progress::{ProgressUpdate, TrackingProgressUpdater};
+
+use crate::{ProgressUpdate, TrackingProgressUpdater};
+
+pub struct FileXorbDependency {
+    pub file_id: u64,
+    pub xorb_hash: MerkleHash,
+    pub n_bytes: u64,
+    pub is_external: bool,
+}
 
 /// A type with with to track a File ID; reporting is done by Arc<str>, but
 /// this ensures the bookkeeping is correct across duplicates and speeds up the
@@ -280,9 +287,10 @@ impl CompletionTracker {
 #[cfg(test)]
 mod tests {
     use merklehash::MerkleHash;
-    use utils::progress::{NoOpProgressUpdater, ProgressUpdaterVerificationWrapper};
 
     use super::*;
+    use crate::verification_wrapper::ProgressUpdaterVerificationWrapper;
+    use crate::NoOpProgressUpdater;
 
     /// A basic test showing partial updates and final completion checks
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
